Move internal/resolver to resolver to make it usable from outside
