+++ resolver/testdata/test-project/db/.envLarge
