+++ resolver/testdata/test-project/invalid.mtaext
