+++ resolver/testdata/test-project/invalid3.mtaext
