+++ resolver/testdata/test-project/mta.yaml
