+++ resolver/testdata/test-project/mtaBad.yaml
