+++ resolver/testdata/test-project/mtaExtTest.yaml
