+++ resolver/testdata/test-project/mtaLargeEnv.yaml
