+++ resolver/testdata/test-project/srv/.env2
