+++ resolver/testdata/test-project/srv/.envExtTest
