+++ resolver/testdata/test-project/srv/.env_with_vcap
