+++ resolver/testdata/test-project/valid1.mtaext
