+++ resolver/testdata/test-project/valid2.mtaext
